refactor all utils
