+++ packages/webpack-config/src/addons/__tests__/withReporting-test.js
@@ -1,6 +1,6 @@
 import path from 'path';
 
-import normalizePaths from '../utils/normalizePaths';
+import { normalizePaths } from '../../utils';
 import withReporting, { throwDeprecatedConfig } from '../withReporting';
 
 const projectRoot = path.resolve(__dirname, '../../../tests/basic');
