+++ packages/webpack-config/src/addons/withWorkbox.ts
@@ -9,8 +9,8 @@ import {
 } from 'workbox-webpack-plugin';
 
 import { AnyConfiguration } from '../types';
-import { isEntry } from '../utils/loaders';
-import { getPaths } from '../utils';
+import { isEntry } from '../utils';
+import { getPaths } from '../env';
 
 export type OfflineOptions = {
   projectRoot?: string;
