+++ packages/webpack-config/src/loaders/__tests__/createAllLoaders-test.js
@@ -1,6 +1,6 @@
 import path from 'path';
 
-import { conditionMatchesFile, getRules } from '../../utils/loaders';
+import { conditionMatchesFile, getRules } from '../../utils/search';
 import createAllLoaders from '../createAllLoaders';
 
 const projectRoot = path.resolve(__dirname, '../../../tests/basic');
