+++ packages/webpack-config/src/utils/index.ts
@@ -1,7 +1,3 @@
-export * from './paths';
-export * from './extensions';
-export { default as getMode } from './getMode';
-export { default as getConfig } from './getConfig';
+export * from './search';
 export { default as normalizePaths } from './normalizePaths';
-export { validateEnvironment } from './validate';
-export { DEFAULT_ALIAS } from './config';
+export * from './config';
