+++ packages/webpack-config/src/utils/search.ts
