+++ packages/webpack-config/withUnimodules.js
@@ -1,4 +1,4 @@
-module.exports = require('./webpack/withUnimodules').default;
+module.exports = require('./addons').withUnimodules;
 
 console.warn(
   '`@expo/webpack-config/withUnimodules` is deprecated, please use `import { withUnimodules } from @expo/webpack-config/addons` instead.'
